nova-assistant/requirements.txt
# Nova Assistant - Dependencies
# Install with: pip install -r requirements.txt

# LLM client (talks to local Ollama server)
ollama>=0.4.0

# Speech-to-Text (offline, GPU-accelerated)
faster-whisper>=1.1.0

# Text-to-Speech (offline, natural voice)
piper-tts>=1.2.0

# Audio capture
sounddevice>=0.5.0
numpy>=1.24.0

# Voice Activity Detection
webrtcvad>=2.0.10

# Hotkey / keyboard listener
keyboard>=0.13.5

# GUI automation
pyautogui>=0.9.54

# Virtual desktop management (Windows)
pyvda>=0.4.0

# System tray icon (future)
# pystray>=0.19.0

# Config
pyyaml>=6.0

# Logging (colorized)
colorama>=0.4.6

# File indexing
whoosh>=2.7.4
